添加微分和积分处理的功能

+++ src/main/java/process/Process.java
@@ -3,5 +3,5 @@ package process;
 import java.util.List;
 
 public interface Process {
-    public List<Short> process(List<Short> list);
+    public List<Integer> process(List<Short> list);
 }
